--- wormhole-infra/wormhole-bus/src/main/java/com/zergclan/wormhole/bus/demo/message/PlanMessage.java
@@ -1,41 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.zergclan.wormhole.bus.demo.message;
-
-/**
- * Plan message.
- */
-public final class PlanMessage {
-
-    /**
-     * Plan desc.
-     */
-    private final String desc;
-
-    public PlanMessage(final String desc) {
-        this.desc = desc;
-    }
-
-    /**
-     * Get desc.
-     * @return Desc.
-     */
-    public String getDesc() {
-        return desc;
-    }
-}
